% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{areas.sp}
\alias{areas.sp}
\title{Six spatial units of analysis}
\format{A SpatialPoylgonsDataFrame with 6 rows and 1 variable:
\describe{
  \item{ID}{ID number of the area}
}}
\source{
\url{https://github.com/nickmalleson/spatialtest}
}
\usage{
areas.sp
}
\description{
A SpatialPolygons object of 6 areas
}
\keyword{datasets}
