% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points1.sp}
\alias{points1.sp}
\title{Base Points dataset: 9 points}
\format{A SpatialPointsDataFrame with 9 rows and 4 variables:
\describe{
  \item{TYPE}{Type of transportation node}
  \item{ROUTENUM}{Associated Route number}
  \item{STATNUM}{Unimportant variable}
  \item{ID}{ID number of the spatial point}
}}
\source{
\url{https://github.com/sppt/sppt/tree/master/java/}
}
\usage{
points1.sp
}
\description{
A SpatialPointsDataFrame object with 9 points
}
\keyword{datasets}
