% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppt_diff.R
\name{sppt_diff}
\alias{sppt_diff}
\title{Performs a Spatial Point Pattern Test following Wheeler et al. (2018)}
\usage{
sppt_diff(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Chi2_Nmin1",
  adj = "BY")
}
\arguments{
\item{p1.sp}{the first points of type SpatialPoints*, order does not matter}

\item{p2.sp}{the second points of type SpatialPoints*, order does not matter}

\item{uoa.sp}{the units of analysis of type SpatialPolygons*}

\item{conf_level}{confidence interval, default = 95}

\item{test}{test to conduct, currently either 'Yates' or 'Fisher', default Yates}

\item{adj}{method to adjust p-values (as per p.adjust), default "BY". To turn off, set to "none".}

\item{tsmethod}{test to conduct, choice of "Chi2_Nmin1", "Chi2",
"Yates", or "Fisher", default = "Chi2_Nmin1"}
}
\value{
Returns uoa.sp (a spatial polygons data frame) including sppt_diff outcomes.
}
\description{
As opposed to the traditional SPPT, this test calculates the difference in
the proportions using either a variety of Chi-square proportions tests or using
Fisher's exact test, and then the p-values are adjusted for multiple comparisons.
See https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3111822 for a description
and example.
}
\examples{
# Plot areas, base points data, and test points data
plot(areas.sp)
text(coordinates(areas.sp), label = areas.sp$ID)
text(coordinates(points1.sp), label = points1.sp$ID, col="blue")
text(coordinates(points2.sp), label = points2.sp$ID, col="red")

output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Chi2_Nmin1")
summary.sppt(output)
output@data

output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Fisher")
summary.sppt(output)
output@data

sppt_diff(points1.sp, points2.sp, areas.sp, test="????") # should get error message

# Vancouver data
set.seed(9866)
myoutput <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp)
summary.sppt(myoutput)

# The sppt_diff() outcomes without correction for multiple comparisons
# are close to sppt_boot() outcomes:
set.seed(9866)
myoutput1 <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, test = "Chi2_Nmin1", adj = "none")
summary.sppt(myoutput1)

set.seed(9866)
myoutput2 <- sppt_boot(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp)
summary.sppt(myoutput2)

}
