% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{summary.globalS.robust}
\alias{summary.globalS.robust}
\title{A convenience function to output the robust global S-value}
\usage{
\method{summary}{globalS.robust}(sppt.sp)
}
\arguments{
\item{sppt.sp}{the spatialobject generated by the \code{sppt} function}
}
\value{
returns the robust global S-value
}
\description{
This function takes as input a spatial object generated by the \code{sppt} function
and returns the robust global S-value: the mean of all robust local \code{similarity}
values. The robust global S-value is the percentage of areas with a similar proportion
of test points as compared to the base points, not counting areas in which no points
occur in the base data and the test data.
}
\examples{
set.seed(76772)
myoutput <- sppt(base_points.sp = points1.sp, test_points.sp = points2.sp, uoa.sp = areas.sp)
summary.globalS.robust(myoutput)

}
\keyword{internal}
