% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppt.R
\name{summary.sppt}
\alias{summary.sppt}
\title{A convenience function to output S-values}
\usage{
summary.sppt(sppt.sp)
}
\arguments{
\item{sppt.sp}{the spatialobject generated by the \code{sppt} function}
}
\value{
returns the standard and robust global S-values
}
\description{
This function takes as input a spatial object generated by the \code{sppt} function
and returns the standard global S-value and the robust global S-value. The standard
global S-value refers to the mean of all local \code{similarity} values. The standard
global S-value is the percentage of areas with a similar proportion
of test points as compared to the base points.
The robust global S-value refers to the mean of all robust local \code{similarity}
values. The robust global S-value is the percentage of areas with a similar proportion
of test points as compared to the base points, not counting areas in which no points
occur in the base data and the test data.
}
\examples{
set.seed(76772)
myoutput <- sppt(base_points.sp = points1.sp, test_points.sp = points2.sp, uoa.sp = areas.sp)
summary.sppt(myoutput)

}
