% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{summary_globalS.generalized}
\alias{summary_globalS.generalized}
\title{A convenience function to output the generalized robust global S-value}
\usage{
summary_globalS.generalized(sppt.sp)
}
\arguments{
\item{sppt.sp}{the spatialobject generated by the \code{sppt} function}
}
\value{
returns the generalized robust global S-value
}
\description{
This function takes as input a spatial object generated by the \code{sppt} function
and returns the generalized robust global S-value.
}
\examples{
# set.seed(76772)
# myoutput <- sppt(base_points.sp = points1.sp, test_points.sp = points2.sp, uoa.sp = areas.sp)
# sppt:::summary_globalS.generalized(myoutput)

}
\keyword{internal}
