% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{summary_globalS.standard}
\alias{summary_globalS.standard}
\title{A convenience function to output the standard global S-value}
\usage{
summary_globalS.standard(sppt.sp)
}
\arguments{
\item{sppt.sp}{the spatialobject generated by the \code{sppt} function}
}
\value{
returns the standard global S-value
}
\description{
This function takes as input a spatial object generated by the \code{sppt} function
and returns the standard global S-value: the mean of all local \code{similarity}
values. The standard global S-value is the percentage of areas with a similar proportion
of test points as compared to the base points.
}
\examples{
# set.seed(76772)
# myoutput <- sppt(base_points.sp = points1.sp, test_points.sp = points2.sp, uoa.sp = areas.sp)
# sppt:::summary_globalS.standard(myoutput)

}
\keyword{internal}
