% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support_functions.R
\name{summary_sppt}
\alias{summary_sppt}
\title{A convenience function to output S-values}
\usage{
summary_sppt(sppt.sp)
}
\arguments{
\item{sppt.sp}{the spatialobject generated by the \code{sppt} function}
}
\value{
returns the standard, robust, and generalized robust global S-values
}
\description{
This function takes as input a spatial object generated by the \code{sppt} function
and returns the standard global S-value, the robust global S-value, and the
generalized robust S-value in a list.
}
\examples{
set.seed(76772)
myoutput <- sppt(base_points.sp = points1.sp, test_points.sp = points2.sp, uoa.sp = areas.sp)
summary_sppt(myoutput)

}
