% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vancouver_areas.sp}
\alias{vancouver_areas.sp}
\title{Vanvouver data: 108 spatial units of analysis}
\format{A SpatialPoylgonsDataFrame with 108 rows and 3 variables:
\describe{
  \item{CTUID}{ID number of the area}
  \item{CMAUID}{Unimportant variable}
  \item{PRUID}{Unimportant variable}
}}
\usage{
vancouver_areas.sp
}
\description{
A SpatialPolygons object of 108 areas
}
\keyword{datasets}
