% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppt_diff.R
\name{sppt_diff}
\alias{sppt_diff}
\title{Performs a Spatial Point Pattern Test (SPPT) following Wheeler et al. (2018)}
\usage{
sppt_diff(p1.sp, p2.sp, uoa.sp, conf_level = 95, test = "Yates",
  tsmethod = "minlike", adj = "BY")
}
\arguments{
\item{p1.sp}{the first points spatialobject, order does not matter}

\item{p2.sp}{the second points spatialobject, order does not matter}

\item{uoa.sp}{the units of analysis spatial object}

\item{conf_level}{confidence interval, default = 95}

\item{test}{test to conduct, currently either 'Yates' or 'Fisher', default Yates}

\item{tsmethod}{if test equals 'Fisher', then exact2x2:exact2x2() is used, with default 'minlike'.
With Yates's test, method is ignored.}

\item{adj}{method to adjust p-values (as per p.adjust), default "BY"}
}
\value{
Returns a spatial polygons data frame object that has fields related to the test
        calculated. When test equals \code{Fisher}, the lower and upper confidence intervals are
        for the odds ratio. When test equals \code{Yates}, the lower and upper confidence intervals
        are for the difference in proportions
}
\description{
As opposed to the traditional SPPT, this test calculates the difference in
the proportions using either a Chi-square proportions test or using Fisher's
exact test, the p-values are then adjusted for multiple comparisons. See
https://papers.ssrn.com/sol3/papers.cfm?abstract_id=3111822 for a description
and example. \cr
If test = 'Fisher', Fisher's exact test is used but with adjusted confidence
intervals (method = 'minlike'). See the R package exact2x2 for details
(https://cran.r-project.org/package=exact2x2)
}
\examples{
# Plot areas, base points data, and test points data
plot(areas.sp)
text(coordinates(areas.sp), label = areas.sp$ID)
text(coordinates(points1.sp), label = points1.sp$ID, col="blue")
text(coordinates(points2.sp), label = points2.sp$ID, col="red")

output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Yates")
summary.sppt(output)
output@data

output <- sppt_diff(points1.sp, points2.sp, areas.sp, test="Fisher")
summary.sppt(output)
output@data

sppt_diff(points1.sp, points2.sp, areas.sp, test="????") # should get error message

# Vancouver data
set.seed(9866)
myoutput <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp)
summary.sppt(myoutput)

# The sppt_diff() outcomes without correction for multiple comparisons
# are very close to sppt_boot() outcomes:
set.seed(9866)
myoutput1 <- sppt_diff(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, test = "Fisher", adj = "none")
summary.sppt(myoutput1)

set.seed(9866)
myoutput2 <- sppt_boot(vancouver_points1.sp, vancouver_points2.sp, vancouver_areas.sp, nsamples = 200)
summary.sppt(myoutput2)

}
